Release Experience Section V2 and UI treatment
