--- src/components/experience/experience-period-short.tsx
@@ -1,21 +1,0 @@
-import { formatYearDate } from '../../utils'
-
-// DEPRICATED: Soon this will become depricated when the V2 of Experience is out
-export default function ExperiencePeriodShort({
-  start,
-  end,
-}: {
-  start: string
-  end: string
-}) {
-  const startDate = new Date(start)
-  const endDate = new Date(end)
-
-  return (
-    <>
-      {formatYearDate(start)}
-      {startDate.getFullYear() !== endDate.getFullYear() &&
-        ` - ${formatYearDate(end)}`}
-    </>
-  )
-}
