+++ src/data/ui.ts
@@ -1,3 +1,3 @@
 import { EXPERIENCE_ID } from '../constants/experience'
 
-export const experienceOpenByDefault = EXPERIENCE_ID.softwareEngineer
+export const experienceOpenByDefault = EXPERIENCE_ID.coFounder
