+++ src/pages/lab/experience/index.tsx
@@ -1,12 +1,12 @@
 import { GetStaticProps } from 'next'
-import ExperienceV2 from '../../../components/experience/experience-v2'
+import Experience from '../../../components/experience'
 import { experiences } from '../../../data/experience'
 import { experienceOpenByDefault } from '../../../data/ui'
 
 export default function ExperiencePage({ experiences, openByDefault }) {
   return (
     <main>
-      <ExperienceV2 experiences={experiences} openByDefault={openByDefault} />
+      <Experience experiences={experiences} openByDefault={openByDefault} />
     </main>
   )
 }
